test(language-server): split lsp flow integration tests
